+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 
 ## [unreleased]
 
+
+ * Portal: Don't log an error if the client aborts before the resource response is being sent (Error: aborted)
+
 ## 2.5.3 (January 22, 2024)
 
  * Remote App Registry: Renamed and added metrics: